// RookieEcommerce.AdminSite/src/api/models/CreateCategoryCommand.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NashLux API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateCategoryCommand
 */
export interface CreateCategoryCommand {
    /**
     * 
     * @type {string}
     * @memberof CreateCategoryCommand
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateCategoryCommand
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CreateCategoryCommand
     */
    parentCategoryId?: string | null;
}

/**
 * Check if a given object implements the CreateCategoryCommand interface.
 */
export function instanceOfCreateCategoryCommand(value: object): value is CreateCategoryCommand {
    return true;
}

export function CreateCategoryCommandFromJSON(json: any): CreateCategoryCommand {
    return CreateCategoryCommandFromJSONTyped(json, false);
}

export function CreateCategoryCommandFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateCategoryCommand {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'parentCategoryId': json['parentCategoryId'] == null ? undefined : json['parentCategoryId'],
    };
}

export function CreateCategoryCommandToJSON(json: any): CreateCategoryCommand {
    return CreateCategoryCommandToJSONTyped(json, false);
}

export function CreateCategoryCommandToJSONTyped(value?: CreateCategoryCommand | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'description': value['description'],
        'parentCategoryId': value['parentCategoryId'],
    };
}