// RookieEcommerce.AdminSite/src/api/models/UpdateCategoryCommand.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NashLux API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateCategoryCommand
 */
export interface UpdateCategoryCommand {
    /**
     * 
     * @type {string}
     * @memberof UpdateCategoryCommand
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateCategoryCommand
     */
    description?: string | null;
}

/**
 * Check if a given object implements the UpdateCategoryCommand interface.
 */
export function instanceOfUpdateCategoryCommand(value: object): value is UpdateCategoryCommand {
    return true;
}

export function UpdateCategoryCommandFromJSON(json: any): UpdateCategoryCommand {
    return UpdateCategoryCommandFromJSONTyped(json, false);
}

export function UpdateCategoryCommandFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateCategoryCommand {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function UpdateCategoryCommandToJSON(json: any): UpdateCategoryCommand {
    return UpdateCategoryCommandToJSONTyped(json, false);
}

export function UpdateCategoryCommandToJSONTyped(value?: UpdateCategoryCommand | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'description': value['description'],
    };
}